+++ be/src/main/java/com/sba301/code/be/service/OrderService.java
@@ -8,8 +8,8 @@ import java.util.List;
 @Service
 public interface OrderService {
     public List<Order> getAllOrders();
-    public Order getOrderById(int orderId);
+    public Order getOrderById(Long orderId);
     public Order createOrder(Order order);
-    public Order updateOrder(int orderId, Order order);
-    public void deleteOrder(int orderId);
+    public Order updateOrder(Long orderId, Order order);
+    public void deleteOrder(Long orderId);
 }
