+++ be/src/main/java/com/sba301/code/be/service/ProductService.java
@@ -8,8 +8,8 @@ import java.util.List;
 @Service
 public interface ProductService {
     public List<Product> getAllProducts();
-    public Product getProductById(int productId);
+    public Product getProductById(Long productId);
     public Product createProduct(Product product);
-    public Product updateProduct(int productId, Product product);
-    public void deleteProduct(int productId);
+    public Product updateProduct(Long productId, Product product);
+    public void deleteProduct(Long productId);
 }
